:bug: Check vtable_ptr before calling destructor
I'm making this PR for documentation purposes. I noticed that a project
of mine that used inplace_function had a vtable_ptr that was set to
nullptr, causing a segfault. I added a check to confirm that the pointer
is not null before running the virtual destructor. Somewhere in the code
the vtable_ptr is not being set correctly. Having these checks
eliminates the issue but doesn't solve the root problem.